requirements.txt
absl-py==0.7.1
astor==0.8.0
audioread==2.1.8
backcall==0.1.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
colorthief==0.2.1
cycler==0.10.0
Cython==0.29.10
decorator==4.4.0
dlib==19.18.0
dominate==2.4.0
easydict==1.9
gast==0.2.2
google-pasta==0.1.7
graphviz==0.8.4
grpcio==1.21.1
h5py==2.9.0
idna==2.8
imageio==2.5.0
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.14.0
joblib==0.13.2
jsonpatch==1.24
jsonpointer==2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
librosa==0.7.0
llvmlite==0.29.0
Markdown==3.1.1
matplotlib==3.1.0
mxnet==1.5.1.post0
mxnet-cu80==1.5.0
networkx==2.3
numba==0.44.1
numpy==1.16.4
olefile==0.46
opencv-contrib-python==4.1.1.26
opencv-python==4.1.0.25
pandas==0.25.1
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==4.3.0
POT==0.4.0
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
pyparsing==2.4.0
pyssim==0.4
python-dateutil==2.8.0
python-speech-features==0.6
pytz==2019.2
PyWavelets==1.0.3
pyzmq==18.1.0
requests==2.22.0
resampy==0.2.1
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.2.0
six==1.12.0
SoundFile==0.10.2
tensorboard==1.14.0
tensorflow-estimator==1.14.0rc1
tensorflow-gpu==1.14.0
termcolor==1.1.0
torch>=1.1.0
torchfile>=0.1.0
torchvision>=0.4.0
tornado==6.0.3
tqdm==4.19.6
traitlets==4.3.2
urllib3==1.25.3
visdom==0.1.8.9
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.4
wrapt==1.11.1
